Split content browser to own window
