+++ src/Display/Menus/TopBar/QDisplay_TopBar.h
@@ -2,6 +2,7 @@
 
 #include <fstream>
 #include <imgui.h>
+#include <memory>
 
 #include "../../../Config/config.h"
 #include "../../../QLogger.h"
@@ -15,6 +16,7 @@
 #include "QDisplay_ImportVAE.h"
 #include "QDisplay_LoadModel.h"
 #include "QDisplay_PluginsWindow.h"
+#include "QDisplay_ContentBrowser.h"
 
 class QDisplay_TopBar : public QDisplay_Base {
 
@@ -24,13 +26,13 @@ private:
   bool logFileOpen = false;
   bool loadFileOpen = false;
   bool selectCanvasOpen = false;
-  bool configureModelOpen = false;
 
   std::unique_ptr<QDisplay_ConfigureModel> m_configureModelWindow;
   std::unique_ptr<QDisplay_ImportModel> m_importModelWindow;
   std::unique_ptr<QDisplay_ImportVAE> m_importVAEWindow;
   std::unique_ptr<QDisplay_LoadModel> m_loadModelWindow;
   std::unique_ptr<QDisplay_PluginsWindow> m_pluginsWindow;
+  std::unique_ptr<QDisplay_ContentBrowser> m_contentBrowserWindow;
 
   // Docker status icons
   std::unique_ptr<Image> m_docker_connected_icon;
@@ -165,6 +167,7 @@ public:
     m_importVAEWindow = std::unique_ptr<QDisplay_ImportVAE>(new QDisplay_ImportVAE(rm, w));
     m_loadModelWindow = std::unique_ptr<QDisplay_LoadModel>(new QDisplay_LoadModel(rm, w));
     m_pluginsWindow = std::unique_ptr<QDisplay_PluginsWindow>(new QDisplay_PluginsWindow(rm, w));
+    m_contentBrowserWindow = std::unique_ptr<QDisplay_ContentBrowser>(new QDisplay_ContentBrowser(rm, w));
 
     // Load images
     m_docker_connected_icon = std::unique_ptr<Image>(new Image(32, 32, "connected_icon"));
@@ -188,6 +191,10 @@ public:
 
         if (ImGui::MenuItem("Load Canvas")) {
           loadFileOpen = true;
+        }
+
+        if (ImGui::MenuItem("Content Browser")) {
+          m_contentBrowserWindow->openWindow();
         }
 
         if (ImGui::MenuItem("Import Model")) {
@@ -266,5 +273,6 @@ public:
     m_importVAEWindow->render();
     m_loadModelWindow->render();
     m_pluginsWindow->render();
+    m_contentBrowserWindow->render();
   }
 };
